refactor: remove prompt barrel and inline single-caller helper
Two small abstraction-discipline cleanups found during an architecture
audit: `src/agent/prompt/index.ts` was a pure re-export barrel with no
documented public surface (banned by the "no convenience barrels" rule),
so its 7 importers now import `PromptBuilder` directly. `executeAgent.ts`'s
`buildWorkflowFlowResult` had exactly one call site, so it's inlined at
that site per the single-caller-extraction rule.

Also documents (without changing behavior) why `leanSession.ts` reads
files via raw `fs/promises` instead of `platform().fs`: it must read the
same on-disk bytes the spawned `lean --server` subprocess sees, which a
host's virtual/faked workspace fs would not provide — confirmed by
`DirectLspAdapter.vitest.ts` regressing when routed through platform().
